DWC tokens replacement fixed
