+++ EventListener/DynamicContentSubscriber.php
@@ -6,57 +6,36 @@ namespace MauticPlugin\CustomObjectsBundle\EventListener;
 
 use Mautic\DynamicContentBundle\DynamicContentEvents;
 use Mautic\DynamicContentBundle\Event\ContactFiltersEvaluateEvent;
-use Mautic\EmailBundle\EventListener\MatchFilterForLeadTrait;
-use MauticPlugin\CustomObjectsBundle\Exception\InvalidArgumentException;
+use Mautic\LeadBundle\Entity\Lead;
+use Mautic\LeadBundle\Entity\Tag;
+use Mautic\LeadBundle\Helper\PrimaryCompanyHelper;
 use MauticPlugin\CustomObjectsBundle\Exception\InvalidSegmentFilterException;
-use MauticPlugin\CustomObjectsBundle\Exception\NotFoundException;
-use MauticPlugin\CustomObjectsBundle\Helper\QueryFilterHelper;
+use MauticPlugin\CustomObjectsBundle\Helper\ContactFilterMatcher;
 use MauticPlugin\CustomObjectsBundle\Provider\ConfigProvider;
-use MauticPlugin\CustomObjectsBundle\Repository\DbalQueryTrait;
 use MauticPlugin\CustomObjectsBundle\Segment\Query\Filter\QueryFilterFactory;
-use PDOException;
-use Psr\Log\LoggerInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
 class DynamicContentSubscriber implements EventSubscriberInterface
 {
-    use MatchFilterForLeadTrait;
-    use DbalQueryTrait;
-
-    /**
-     * @var QueryFilterFactory
-     */
-    private $queryFilterFactory;
-
-    /**
-     * @var QueryFilterHelper
-     */
-    private $queryFilterHelper;
-
-    /**
-     * @var ConfigProvider
-     */
-    private $configProvider;
-
-    /**
-     * @var LoggerInterface
-     */
-    private $logger;
+    private QueryFilterFactory $queryFilterFactory;
+    private ConfigProvider $configProvider;
+    private PrimaryCompanyHelper $primaryCompanyHelper;
+    private ContactFilterMatcher $contactFilterMatcher;
 
     public function __construct(
         QueryFilterFactory $queryFilterFactory,
-        QueryFilterHelper $queryFilterHelper,
         ConfigProvider $configProvider,
-        LoggerInterface $logger
+        PrimaryCompanyHelper $primaryCompanyHelper,
+        ContactFilterMatcher $contactFilterMatcher,
     ) {
-        $this->queryFilterFactory = $queryFilterFactory;
-        $this->queryFilterHelper  = $queryFilterHelper;
-        $this->configProvider     = $configProvider;
-        $this->logger             = $logger;
+        $this->queryFilterFactory   = $queryFilterFactory;
+        $this->configProvider       = $configProvider;
+        $this->primaryCompanyHelper = $primaryCompanyHelper;
+        $this->contactFilterMatcher = $contactFilterMatcher;
     }
 
     /**
-     * @return mixed[]
+     * @return array<string,array{string,int}>
      */
     public static function getSubscribedEvents(): array
     {
@@ -65,47 +44,52 @@ class DynamicContentSubscriber implements EventSubscriberInterface
         ];
     }
 
-    /**
-     * @throws InvalidArgumentException
-     * @throws NotFoundException
-     */
     public function evaluateFilters(ContactFiltersEvaluateEvent $event): void
     {
-        if (!$this->configProvider->pluginIsEnabled()) {
+        if ($event->isEvaluated()
+            || !$this->configProvider->pluginIsEnabled()
+            || !$this->hasCustomObjectFilters($event->getFilters())
+        ) {
             return;
         }
 
-        $eventFilters = $event->getFilters();
+        $event->setIsEvaluated(true);
+        $event->stopPropagation();
+        $event->setIsMatched($this->doesFiltersMatch($event->getFilters(), $event->getContact()));
+    }
 
-        if ($event->isEvaluated()) {
-            return;
+    /**
+     * @param mixed[] $filters
+     */
+    private function hasCustomObjectFilters(array $filters): bool
+    {
+        foreach ($filters as $filter) {
+            try {
+                $this->queryFilterFactory->configureQueryBuilderFromSegmentFilter($filter, 'filter');
+
+                return true;
+            } catch (InvalidSegmentFilterException) {
+            }
         }
 
-        foreach ($eventFilters as $key => $eventFilter) {
-            $queryAlias = "filter_{$key}";
+        return false;
+    }
 
-            try {
-                $filterQueryBuilder = $this->queryFilterFactory->configureQueryBuilderFromSegmentFilter($eventFilter, $queryAlias);
-            } catch (InvalidSegmentFilterException $e) {
-                continue;
-            }
+    /**
+     * @param mixed[] $filters
+     */
+    private function doesFiltersMatch(array $filters, Lead $contact): bool
+    {
+        $lead = $this->primaryCompanyHelper->getProfileFieldsWithPrimaryCompany($contact);
+        $lead = array_merge($lead, [
+            'tags' => array_map(
+                function (Tag $v) {
+                    return $v->getId();
+                },
+                $contact->getTags()->toArray()
+            ),
+        ]);
 
-            $this->queryFilterHelper->addContactIdRestriction($filterQueryBuilder, $queryAlias, (int) $event->getContact()->getId());
-
-            try {
-                if ($this->executeSelect($filterQueryBuilder)->rowCount()) {
-                    $event->setIsEvaluated(true);
-                    $event->setIsMatched(true);
-                } else {
-                    $event->setIsEvaluated(true);
-                }
-            } catch (PDOException $e) {
-                $this->logger->addError('Failed to evaluate dynamic content for custom object '.$e->getMessage());
-
-                throw $e;
-            }
-
-            $event->stopPropagation();  // The filter is ours, we won't allow no more processing
-        }
+        return $this->contactFilterMatcher->match($filters, $lead);
     }
 }
